+++ src/traits/i-history/CHANGELOG.md
@@ -9,6 +9,12 @@ Changelog
 > - :house:      [Internal]
 > - :nail_care:  [Polish]
 
+## v3.0.0-rc.??? (2020-12-??)
+
+#### :bug: Bug Fix
+
+* Fixed an issue when the next calling of `initIndex` pushes another one index page
+
 ## v3.0.0-rc.49 (2020-08-03)
 
 #### :bug: Bug Fix